% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts_WhyR2017.R
\docType{data}
\name{abstracts_WhyR2017}
\alias{abstracts_WhyR2017}
\title{WhyR? 2017 abstracts}
\format{a data frame with 47 rows and 4 columns}
\usage{
data(abstracts_WhyR2017)
}
\description{
Abstracts from the WhyR? 2017 Conference. Abstracts are in Polish.
}
\details{
\itemize{
\item author
\item affiliation
\item title
\item abstract
}
}
\keyword{abstracts}
