% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts_WhyR2018.R
\docType{data}
\name{abstracts_WhyR2018}
\alias{abstracts_WhyR2018}
\title{WhyR? 2018 abstracts}
\format{a data frame with 30 rows and 4 columns}
\usage{
data(abstracts_WhyR2018)
}
\description{
Abstracts from the WhyR? 2018 Conference. Abstracts are in English.
}
\details{
\itemize{
\item author
\item affiliation
\item title
\item abstract
}
}
\keyword{abstracts}
